+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * Now avoids `libc` for windows builds, where it was unused
  * Grouped codes into sub-modules
  * Changed codes from statics to constants
+ * Added and used a macro for creating sequences
 
 # 0.1 (January 24th, 2018)